Moved `oris.domain` package to `oris.application.apiclient.dto`, added new DTO classes, refactored `oris.application` components, updated all references, and upgraded test dependencies.
