+++ backend/build.gradle.kts
@@ -83,8 +83,8 @@ dependencies {
     //implementation("org.jmolecules:jmolecules-hexagonal-architecture:1.9.0")
     implementation("org.jmolecules:jmolecules-onion-architecture:1.9.0")
     implementation("org.jmolecules:jmolecules-events:1.9.0")
-    testImplementation("org.jmolecules.integrations:jmolecules-archunit:0.20.0")
-    testImplementation("com.tngtech.archunit:archunit-junit5:1.3.0")
+    testImplementation("org.jmolecules.integrations:jmolecules-archunit:0.28.0")
+    testImplementation("com.tngtech.archunit:archunit-junit5:1.4.1")
 }
 
 tasks.withType<Test> {
