+++ backend/src/main/java/club/klabis/oris/application/apiclient/OrisApiClient.java
@@ -1,17 +1,19 @@
-package club.klabis.oris.adapters.apiclient;
+package club.klabis.oris.application.apiclient;
 
 
-import club.klabis.oris.domain.*;
+import club.klabis.oris.application.apiclient.dto.OrisEvent;
+import club.klabis.oris.application.apiclient.dto.OrisEventListFilter;
+import club.klabis.oris.application.apiclient.dto.OrisUserInfo;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.service.annotation.GetExchange;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Map;
 
+// TODO: move into adapters
 public interface OrisApiClient {
 
     int CLUB_ID_ZBM = 205;
@@ -35,39 +37,6 @@ public interface OrisApiClient {
             @JsonProperty("Method") String method
     ) {
 
-    }
-
-    record OrisUserInfo(
-            @JsonProperty("ID")
-            int orisId,
-            @JsonProperty("FirstName")
-            String firstName,
-            @JsonProperty("LastName")
-            String lastName) {
-    }
-
-    record OrisEventListFilter(
-            String region,
-            LocalDate dateFrom,
-            LocalDate dateTo
-    ) {
-        public static OrisEventListFilter EMPTY = new OrisEventListFilter(null, null, null);
-
-        public static OrisEventListFilter createDefault() {
-            return new OrisEventListFilter(OrisApiClient.REGION_JIHOMORAVSKA, LocalDate.now(), LocalDate.now().plusMonths(3));
-        }
-
-        public OrisEventListFilter withRegion(String region) {
-            return new OrisEventListFilter(region, dateFrom, dateTo);
-        }
-
-        public OrisEventListFilter withDateTo(LocalDate dateTo) {
-            return new OrisEventListFilter(region, dateFrom, dateTo);
-        }
-
-        public OrisEventListFilter withDateFrom(LocalDate dateFrom) {
-            return new OrisEventListFilter(region, dateFrom, dateTo);
-        }
     }
 
 }
