+++ backend/src/main/java/club/klabis/oris/application/apiclient/dto/OrisEvent.java
@@ -1,4 +1,4 @@
-package club.klabis.oris.domain;
+package club.klabis.oris.application.apiclient.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
