+++ backend/src/main/java/club/klabis/oris/application/apiclient/dto/OrisEventDiscipline.java
@@ -1,17 +1,15 @@
-package club.klabis.oris.domain;
+package club.klabis.oris.application.apiclient.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record OrisEventSport(
+public record OrisEventDiscipline(
         @JsonProperty("ID")
         int id,
+        @JsonProperty("ShortName")
+        String shortName,
         @JsonProperty("NameCZ")
         String nameCZ,
         @JsonProperty("NameEN")
         String nameEN
 ) {
-    public static int ID_OB = 1;
-    public static int ID_LOB = 2;
-    public static int ID_MTBO = 3;
-    public static int ID_TRAIL = 4;
 }
