--- backend/src/main/java/club/klabis/oris/domain/OrisEventLevel.java
@@ -1,16 +1,0 @@
-package club.klabis.oris.domain;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record OrisEventLevel(
-        @JsonProperty("ID")
-        int id,
-        @JsonProperty("ShortName")
-        String shortName,
-        @JsonProperty("NameCZ")
-        String nameCZ,
-        @JsonProperty("NameEN")
-        String nameEN
-) {
-
-}
