+++ klabis-api-spec.yaml
@@ -967,9 +967,9 @@ components:
         Additional validations:
         - either contact or at least 1 guardian needs to be entered
       x-class-extra-annotation: |-
-        @club.klabis.domain.members.forms.AtLeastOneContactIsDefined.List({
-          @club.klabis.domain.members.forms.AtLeastOneContactIsDefined(contactType = club.klabis.domain.members.Contact.Type.EMAIL, message = "At least one email contact must be provided"),
-          @club.klabis.domain.members.forms.AtLeastOneContactIsDefined(contactType = club.klabis.domain.members.Contact.Type.PHONE, message = "At least one phone contact must be provided")
+        @club.klabis.members.domain.forms.AtLeastOneContactIsDefined.List({
+          @club.klabis.members.domain.forms.AtLeastOneContactIsDefined(contactType = club.klabis.members.domain.Contact.Type.EMAIL, message = "At least one email contact must be provided"),
+          @club.klabis.members.domain.forms.AtLeastOneContactIsDefined(contactType = club.klabis.members.domain.Contact.Type.PHONE, message = "At least one phone contact must be provided")
         })
       required:
         - nationality
